requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.2
openpyxl==3.1.2
newspaper3k==0.2.8
nltk==3.8.1
feedparser==6.0.10
googlesearch-python==1.2.3
gunicorn==21.2.0
lxml_html_clean==0.4.3
transformers==4.36.2
torch==2.2.1
sentencepiece==0.1.99
